Time: 2 ms (100.00%) | Memory: 54.3 MB (73.19%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -1,8 +1,8 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
 
-            double sum = 0.0;
-           double ans = 0;
+            int sum = 0;
+           int ans = 0;
        for(int i=0; i<k; i++){
            sum += nums[i];
        }
@@ -16,6 +16,6 @@ class Solution {
         }
            
        
-        return ans/k;
+        return (double)ans/k;
     }
 }
